move ddtest viewer to apps/
