+++ package.json
@@ -25,8 +25,8 @@
     "serve-language-workbench": "cd apps/languageWorkbench/public && http-server -c-1",
     "build-dl-parser-playground": "esbuild apps/dlParser/main.tsx --bundle --outfile=apps/dlParser/public/bundle.js '--define:process.env.NODE_ENV=\"development\"' --loader:.dl=text --sourcemap --external:path --external:fs --external:crypto --loader:.dl=text --loader:.css=text --loader:.grammar=text",
     "serve-dl-parser-playground": "cd apps/dlParser/public && http-server -c-1",
-    "build-ddtest-viewer": "esbuild util/ddTest/viewer/main.tsx --bundle --outfile=util/ddTest/viewer/public/bundle.js '--define:process.env.NODE_ENV=\"development\"' --sourcemap --external:path --external:fs --external:crypto",
-    "serve-ddtest-viewer": "cd util/ddTest/viewer/public && http-server -c-1",
+    "build-ddtest-viewer": "esbuild apps/ddTestViewer/main.tsx --bundle --outfile=apps/ddTestViewer/public/bundle.js '--define:process.env.NODE_ENV=\"development\"' --sourcemap --external:path --external:fs --external:crypto",
+    "serve-ddtest-viewer": "cd apps/ddTestViewer/public && http-server -c-1",
     "build-sim": "esbuild apps/sim/main.tsx --bundle --outfile=apps/sim/public/bundle.js '--define:process.env.NODE_ENV=\"development\"' --sourcemap --external:path --external:fs --external:crypto --loader:.dl=text",
     "serve-sim": "cd apps/sim/public && http-server -c-1",
     "build-rel-sql-playground": "esbuild apps/relSQLPlayground/main.tsx --bundle --outfile=apps/relSQLPlayground/public/bundle.js '--define:process.env.NODE_ENV=\"development\"' --sourcemap --external:path --external:fs --external:crypto --loader:.dl=text --loader:.grammar=text --loader:.ttf=text",
